requirements.txt
asgiref==3.4.1
Django==3.2.8
django-ckeditor==6.1.0
django-extensions==3.1.3
django-js-asset==1.2.2
Pillow==8.3.2
pytz==2021.3
sqlparse==0.4.2
